+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## 10.6.0 (2025-08-01)
+## 10.6.0 (2025-08-04)
 
 * CONRAT-40669: Add setPageId public interface to generate unique page id (pgid)